# doc/openapi.yml
openapi: 3.0.3
info:
  title: Library Management API
  description: |
    Library management API — manages books, copies, authors, genres,
    arrivals (stock receipts) and sales within one or more libraries.

    Based on the project's conceptual data model (MCD).
  version: 0.1.0
  contact:
    name: API Support
    email: support@library.app
servers:
  - url: https://api.library.app/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
components:
  schemas:
...
    LibraryRequest:
      type: object
      required:
        - name
        - phone
        - email
        - address
      properties:
        name:
          type: string
          maxLength: 100
          example: Librairie Générale
        phone:
          type: string
          maxLength: 30
          example: +261 34 12 345 67
        email:
          type: string
          format: email
          maxLength: 100
          example: contact@librairie-generale.mg
        address:
          type: string
          maxLength: 100
          example: 15 Avenue de l'Indépendance, Antananarivo
    LibraryResponse:
      type: object
      required:
        - id
        - name
        - phone
        - email
        - address
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          maxLength: 100
          example: Librairie Générale
        phone:
          type: string
          maxLength: 30
          example: +261 34 12 345 67
        email:
          type: string
          format: email
          maxLength: 100
          example: contact@librairie-generale.mg
        address:
          type: string
          maxLength: 100
          example: 15 Avenue de l'Indépendance, Antananarivo
    LibraryUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          maxLength: 100
          example: Librairie Générale
        phone:
          type: string
          maxLength: 30
          example: +261 34 12 345 67
        email:
          type: string
          format: email
          maxLength: 100
          example: contact@librairie-generale.mg
        address:
          type: string
          maxLength: 100
          example: 15 Avenue de l'Indépendance, Antananarivo

    # ────────────────── Book ──────────────────
...
    AuthorRequest:
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
          maxLength: 100
          example: Antoine
        lastName:
          type: string
          maxLength: 100
          example: de Saint-Exupéry
    AuthorResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
      properties:
        id:
          type: string
          format: uuid
          example: c1d2e3f4-a5b6-7890-abcd-ef1234567890
        firstName:
          type: string
          maxLength: 100
          example: Antoine
        lastName:
          type: string
          maxLength: 100
          example: de Saint-Exupéry
    AuthorUpdateRequest:
      type: object
      minProperties: 1
      properties:
        firstName:
          type: string
          maxLength: 100
          example: Antoine
        lastName:
          type: string
          maxLength: 100
          example: de Saint-Exupéry

    # ────────────────── Genre ──────────────────
    GenreRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          example: Fiction
    GenreResponse:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: d1e2f3a4-b5c6-7890-abcd-ef1234567890
        name:
          type: string
          maxLength: 100
          example: Fiction
        createdAt:
          type: string
          format: date-time
          example: "2026-06-01T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-06-01T10:30:00Z"
...
    GenreUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          maxLength: 100
          example: Fiction

    # ────────────────── User ──────────────────
...
    Pagination:
      type: object
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 42

    # ────────────────── Shared error ──────────────────
    Error:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field affected by the error
              message:
                type: string
                description: Error description for this field
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login or POST /auth/register
  responses:
    BadRequest:
      description: Malformed request or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 400
            error: Bad Request
            message: The request body contains invalid JSON or a parameter is malformed
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 404
            error: Not Found
            message: The requested resource was not found
    Unprocessable:
      description: Semantically invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 422
            error: Unprocessable Entity
            message: At least one required field is missing or invalid
...
    Unauthorized:
      description: Authentication failed or missing/invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 401
            error: Unauthorized
            message: Invalid or expired JWT token
    Forbidden:
      description: Authenticated but insufficient permissions (requires ADMIN role)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 403
            error: Forbidden
            message: Admin access required
paths:
...
  /libraries:
    get:
      operationId: listLibraries
      summary: List all libraries
      tags:
        - Libraries
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Free-text search on name, email, phone, or address
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Paginated list of libraries
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LibraryResponse'
                  meta:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  -
                    id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                    name: Librairie Générale
                    phone: +261 34 12 345 67
                    email: contact@librairie-generale.mg
                    address: "15 Avenue de l'Indépendance, Antananarivo"
                meta:
                  page: 1
                  size: 20
                  total: 1

        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'

    post:
      operationId: createLibrary
      summary: Create a new library
      tags:
        - Libraries
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryRequest'
      responses:
        '201':
          description: Library created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
              example:
                id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                name: Librairie Générale
                phone: +261 34 12 345 67
                email: contact@librairie-generale.mg
                address: "15 Avenue de l'Indépendance, Antananarivo"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /libraries/{libraryId}:
    get:
      operationId: getLibrary
      summary: Get a library by ID
      tags:
        - Libraries
      parameters:
        - name: libraryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Library details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
              example:
                id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                name: Librairie Générale
                phone: +261 34 12 345 67
                email: contact@librairie-generale.mg
                address: "15 Avenue de l'Indépendance, Antananarivo"

        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateLibrary
      summary: Update a library
      tags:
        - Libraries
      security:
        - BearerAuth: []
      parameters:
        - name: libraryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryUpdateRequest'
      responses:
        '200':
          description: Library updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
              example:
                id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                name: Librairie Générale
                phone: +261 34 12 345 67
                email: contact@librairie-generale.mg
                address: "15 Avenue de l'Indépendance, Antananarivo"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      operationId: deleteLibrary
      summary: Delete a library
      tags:
        - Libraries
      security:
        - BearerAuth: []
      parameters:
        - name: libraryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '204':
          description: Library deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════ Books ═══════════════════════
...
  /authors:
    get:
      operationId: listAuthors
      summary: List all authors
      tags:
        - Authors
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Free-text search on first name or last name
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Paginated list of authors
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuthorResponse'
                  meta:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  -
                    id: c1d2e3f4-a5b6-7890-abcd-ef1234567890
                    firstName: Antoine
                    lastName: de Saint-Exupéry
                meta:
                  page: 1
                  size: 20
                  total: 1

        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createAuthor
      summary: Create a new author
      tags:
        - Authors
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: c1d2e3f4-a5b6-7890-abcd-ef1234567890
                firstName: Antoine
                lastName: de Saint-Exupéry

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /authors/{authorId}:
    get:
      operationId: getAuthor
      summary: Get an author by ID
      tags:
        - Authors
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: c1d2e3f4-a5b6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Author details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: c1d2e3f4-a5b6-7890-abcd-ef1234567890
                firstName: Antoine
                lastName: de Saint-Exupéry

        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateAuthor
      summary: Update an author
      tags:
        - Authors
      security:
        - BearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: c1d2e3f4-a5b6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdateRequest'
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: c1d2e3f4-a5b6-7890-abcd-ef1234567890
                firstName: Antoine
                lastName: de Saint-Exupéry

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      operationId: deleteAuthor
      summary: Delete an author
      tags:
        - Authors
      security:
        - BearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: c1d2e3f4-a5b6-7890-abcd-ef1234567890
      responses:
        '204':
          description: Author deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════ Genres ═══════════════════════
  /genres:
    get:
      operationId: listGenres
      summary: List all genres
      tags:
        - Genres
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Free-text search on name
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Paginated list of genres
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GenreResponse'
                  meta:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  -
                    id: d1e2f3a4-b5c6-7890-abcd-ef1234567890
                    name: Fiction
                    createdAt: "2026-06-01T10:30:00Z"
                    updatedAt: "2026-06-01T10:30:00Z"
                meta:
                  page: 1
                  size: 20
                  total: 1

        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createGenre
      summary: Create a new genre
      tags:
        - Genres
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreRequest'
      responses:
        '201':
          description: Genre created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
              example:
                id: d1e2f3a4-b5c6-7890-abcd-ef1234567890
                name: Fiction
                createdAt: "2026-06-01T10:30:00Z"
                updatedAt: "2026-06-01T10:30:00Z"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /genres/{genreId}:
    get:
      operationId: getGenre
      summary: Get a genre by ID
      tags:
        - Genres
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: d1e2f3a4-b5c6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Genre details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
              example:
                id: d1e2f3a4-b5c6-7890-abcd-ef1234567890
                name: Fiction
                createdAt: "2026-06-01T10:30:00Z"
                updatedAt: "2026-06-01T10:30:00Z"

        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateGenre
      summary: Update a genre
      tags:
        - Genres
      security:
        - BearerAuth: []
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: d1e2f3a4-b5c6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreUpdateRequest'
      responses:
        '200':
          description: Genre updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
              example:
                id: d1e2f3a4-b5c6-7890-abcd-ef1234567890
                name: Fiction
                createdAt: "2026-06-01T10:30:00Z"
                updatedAt: "2026-06-01T10:30:00Z"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      operationId: deleteGenre
      summary: Delete a genre
      tags:
        - Genres
      security:
        - BearerAuth: []
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: d1e2f3a4-b5c6-7890-abcd-ef1234567890
      responses:
        '204':
          description: Genre deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════ Users ═══════════════════════
...